Add few tests and fix bug in builder self.parent is not node.parent
